test: add integration tests for poisson_race simulator behavior overrides
